src/NeedleWork.Application/Features/Users/Queries/GetById/GetUserByIdQueryHandler.cs
using MediatR;
using NeedleWork.Application.ViewModels.Users;
using NeedleWork.Core.Entities;
using NeedleWork.Core.Exceptions;
using NeedleWork.Core.Repositories;
using NeedleWork.Core.Shared;

namespace NeedleWork.Application.Features.Users.Queries.GetById;

public class GetUserByIdQueryHandler : IRequestHandler<GetUserByIdQuery, UserDetailsViewModel>
{
    private readonly IUserRepository _userRepository;

    public GetUserByIdQueryHandler(IUserRepository userRepository)
    {
        _userRepository = userRepository;
    }

    public async Task<UserDetailsViewModel> Handle(GetUserByIdQuery request, CancellationToken cancellationToken)
    {
        User? user = await _userRepository.GetByIdAsync(request.Id);

        if (user is null)
            throw new NotFoundException(Errors.UserNotFound(request.Id));
        
        return new UserDetailsViewModel(
            user.Id,
            user.FullName,
            user.Email,
            user.BirthDate
        );
    }
}